refactor(Car): InputValidator의 validateCarName 함수를 Car 클래스의 생성자로 이동
validateCarName 함수를 Car 클래스의 생성자로 이동하여 검증을 진행한다.
