+++ src/main/java/utils/InputValidator.java
@@ -1,14 +1,6 @@
 package utils;
 
 public class InputValidator {
-    private static final int CAR_NAME_MAX_LENGTH = 5;
-
-    public static void carNamesValidator(String carNames) {
-        for (String carName : carNames.split(",")) {
-            validateCarNameLength(carName);
-            validateCarNameExists(carName);
-        }
-    }
 
     public static void tryNumberValidator(String number) {
         int validateNumber = validateNumber(number);
@@ -16,17 +8,7 @@ public class InputValidator {
     }
 
 
-    private static void validateCarNameLength(String carName) {
-        if (carName.length() > CAR_NAME_MAX_LENGTH) {
-            throw new IllegalArgumentException(ErrorMessage.CAR_NAME_TOO_LONG);
-        }
-    }
 
-    private static void validateCarNameExists(String carName) {
-        if(carName.isBlank()){
-            throw new IllegalArgumentException(ErrorMessage.CAR_NAME_IS_BLANK);
-        }
-    }
 
     private static int validateNumber(String number) {
         try {
